strict ANSI-C conformance implemented in core library
to activate use:

./configure --disable-all-bindings CFLAGS="-ansi -pedantic"

So only useful when no bindings are required really

All xraylib releases should be tested for this to work!

Many thanks to Oleg Shirokobrod of Oxford Instruments for providing the
necessary patches!

+++ src/fluor_lines.c
@@ -13,13 +13,13 @@ THIS SOFTWARE IS PROVIDED BY Bruno Golosio, Antonio Brunetti, Manuel Sanchez del
 
 #include "xrayglob.h"
 #include "xraylib.h"
-#define KL1 -KL1_LINE-1
-#define KL2 -KL2_LINE-1
-#define KL3 -KL3_LINE-1
-#define KM1 -KM1_LINE-1
-#define KM2 -KM2_LINE-1
-#define KM3 -KM3_LINE-1
-#define KP5 -KP5_LINE-1
+#define KL1 -(int)KL1_LINE-1
+#define KL2 -(int)KL2_LINE-1
+#define KL3 -(int)KL3_LINE-1
+#define KM1 -(int)KM1_LINE-1
+#define KM2 -(int)KM2_LINE-1
+#define KM3 -(int)KM3_LINE-1
+#define KP5 -(int)KP5_LINE-1
 
 /*////////////////////////////////////////////////////////////////////
 //                                                                  //
